Move more code to math.eval.list.eol
