+++ mathics/core/systemsymbols.py
@@ -31,6 +31,7 @@ SymbolAll = Symbol("System`All")
 SymbolAlternatives = Symbol("System`Alternatives")
 SymbolAnd = Symbol("System`And")
 SymbolAppend = Symbol("System`Append")
+SymbolAppendTo = Symbol("System`AppendTo")
 SymbolApply = Symbol("System`Apply")
 SymbolArcCos = Symbol("System`ArcCos")
 SymbolArcSin = Symbol("System`ArcSin")
@@ -80,6 +81,7 @@ SymbolDirectedInfinity = Symbol("System`DirectedInfinity")
 SymbolDispatch = Symbol("System`Dispatch")
 SymbolDot = Symbol("System`Dot")
 SymbolDownValues = Symbol("System`DownValues")
+SymbolDrop = Symbol("System`Drop")
 SymbolE = Symbol("System`E")
 SymbolEdgeForm = Symbol("System`EdgeForm")
 SymbolEndOfFile = Symbol("System`EndOfFile")
@@ -267,6 +269,7 @@ SymbolSubscriptBox = Symbol("System`SubscriptBox")
 SymbolSubsuperscriptBox = Symbol("System`SubsuperscriptBox")
 SymbolSuperscriptBox = Symbol("System`SuperscriptBox")
 SymbolTable = Symbol("System`Table")
+SymbolTake = Symbol("System`Take")
 SymbolTan = Symbol("System`Tan")
 SymbolTanh = Symbol("System`Tanh")
 SymbolTeXForm = Symbol("System`TeXForm")
